pyproject.toml
[tool.poetry]
name = "agentic-workflow"
version = "0.1.0"
description = ""
authors = ["Balu Sekar <balu.sekar@fourkites.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.115.11"}
langgraph = "^0.3.16"
langchain-openai = "^0.3.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"